PMM-6863 Move to common
