+++ packages.dot
@@ -1,6 +1,5 @@
 digraph packages {
 	"/service/cluster" -> "/service/k8sclient";
-	"/service/k8sclient" -> "/k8s_api/apimachinery/pkg/apis/meta/v1";
 	"/service/k8sclient" -> "/k8s_api/common";
 	"/service/k8sclient" -> "/k8s_api/psmdb/v1";
 	"/service/k8sclient" -> "/k8s_api/pxc/v1";
